backend/requirements.txt
openai~=0.27.7
fastapi~=0.95.2
gTTS~=2.3.2
python-multipart~=0.0.6
python-jose[cryptography]
ffmpeg-python~=0.2.0
pydantic~=1.10.7
uvicorn~=0.22.0
pydub~=0.25.1
requests~=2.31.0
edge-tts~=6.1.5
elevenlabs~=0.2.15
langchain
llama_index
supabase 
notion-client 
pgvector 
postgrest 
psycopg 
psycopg2-binary 
pydantic 
vecs
google-api-core==2.11.0
google-api-python-client==2.86.0
docarray==0.21.0
docutils==0.20.1
docx2txt==0.8
faiss-cpu==1.7.4
gpt4all